libuser fns and assigning syscallvec

+++ libuser.h
@@ -9,13 +9,13 @@
 extern int  Spawn(char *name, int (*func)(char *), char *arg, long stack_size,
                   long priority, int *pid);
 extern int  Wait(int *pid, int *status);
-extern void Terminate(int status);
+extern void Terminate(long status);
 extern void GetTimeofDay(int *tod);
 extern void CPUTime(int *cpu);
 extern void GetPID(int *pid);
-extern int  SemCreate(int value, int *semaphore);
-extern int  SemP(int semaphore);
-extern int  SemV(int semaphore);
-extern int  SemFree(int semaphore);
+extern int  SemCreate(long value, int *semaphore);
+extern int  SemP(long semaphore);
+extern int  SemV(long semaphore);
+extern int  SemFree(long semaphore);
 
 #endif
